SDESK-5048 Rearrange tests and keep test files consistent
