+++ tests/prepopulate/app_initialization_test.py
@@ -4,8 +4,8 @@ import tempfile
 from flask import json
 from unittest.mock import patch
 
-from .app_initialize import AppInitializeWithDataCommand
-from .app_scaffold_data import AppScaffoldDataCommand
+from apps.prepopulate.app_initialize import AppInitializeWithDataCommand
+from apps.prepopulate.app_scaffold_data import AppScaffoldDataCommand
 from apps.prepopulate.app_initialize import fillEnvironmentVariables
 from superdesk import get_resource_service, app
 from superdesk.tests import TestCase
