+++ tests/rules/routing_rules_tests.py
@@ -1,49 +1,17 @@
-
 import unittest
 from unittest import mock
-
 from copy import deepcopy
-from datetime import datetime, timedelta
+from datetime import datetime
 
-import superdesk
-from superdesk.tests import TestCase
-from .routing_rules import Weekdays
-
-
-class WeekdaysTestCase(unittest.TestCase):
-
-    def test_is_valid_schedule(self):
-        self.assertTrue(Weekdays.is_valid_schedule(['MON', 'TUE']))
-        self.assertFalse(Weekdays.is_valid_schedule(['MON', 'SRC']))
-
-    def test_is_scheduled(self):
-        now = datetime.now()
-        day = now.strftime('%A')[:3]
-
-        self.assertFalse(Weekdays.is_scheduled_day(now, []))
-        self.assertTrue(Weekdays.is_scheduled_day(now, [day]))
-        self.assertFalse(Weekdays.is_scheduled_day(now + timedelta(days=1), [day]))
-
-    def test_weekday_dayname(self):
-        now = datetime.now()
-        day = now.strftime('%A')[:3].upper()
-        self.assertEqual(day, Weekdays.dayname(now))
+from apps.rules.routing_rules import RoutingRuleSchemeService
+from superdesk.errors import SuperdeskApiError
 
 
 class RoutingRuleSchemeServiceTest(unittest.TestCase):
     """Base class for RoutingRuleSchemeService tests."""
 
     def setUp(self):
-        try:
-            from .routing_rules import RoutingRuleSchemeService
-        except ImportError:
-            # a missing class should result in a failed test, not in an error
-            # that prevents all the tests in the module from even being run
-            self.fail(
-                "Could not import class under test "
-                "(RoutingRuleSchemeService).")
-        else:
-            self.instance = RoutingRuleSchemeService()
+        self.instance = RoutingRuleSchemeService()
 
 
 @mock.patch(
@@ -170,7 +138,7 @@ class ValidateScheduleMethodTestCase(RoutingRuleSchemeServiceTest):
         {'Foo/Bar', 'Foo/Baz', 'Here/There'}
     )
     def test_raises_error_on_unknown_time_zone(self):
-        from superdesk.errors import SuperdeskApiError
+
 
         self.schedule['time_zone'] = 'Invalid/Zone'
 
@@ -347,47 +315,3 @@ class GetScheduledRoutingRulesMethodTestCase(RoutingRuleSchemeServiceTest):
         now = datetime(2015, 9, 15, 23, 59, 59, 999999)  # Tuesday
         result = self.instance._get_scheduled_routing_rules(rules, now)
         self.assertEqual(result, [])
-
-
-class SetDefaultValuesTestCase(TestCase):
-    def test_setting_default_values(self):
-
-        from .routing_rules import RoutingRuleSchemeService
-        instance = RoutingRuleSchemeService()
-        category = [{'qcode': 'a', 'name': 'Australian General News'}]
-        result = instance._assign_default_values({'anpa_category': category}, None)
-        self.assertEqual(result['anpa_category'], [{'qcode': 'a', 'name': 'Australian General News'}])
-
-        with self.app.app_context():
-            result = instance._assign_default_values({}, None)
-            self.assertIsNone(result['anpa_category'])
-
-            result = instance._assign_default_values({}, [{'qcode': 'a', 'name': 'Australian General News'}])
-            self.assertEqual(result['anpa_category'], [{'qcode': 'a', 'name': 'Australian General News'}])
-
-    def test_getting_selected_categories(self):
-        vocabularies = [{'_id': 'categories', 'items': [
-            {'qcode': 'a', 'name': 'foo', 'is_active': True},
-            {'qcode': 'b', 'name': 'bar', 'is_active': True},
-            {'qcode': 'c', 'name': 'baz', 'is_active': False},
-        ]}]
-
-        from .routing_rules import RoutingRuleSchemeService
-        instance = RoutingRuleSchemeService()
-
-        with self.app.app_context():
-            self.app.data.insert('vocabularies', vocabularies)
-            result = instance._get_categories('a')
-            self.assertEqual(result, [{'qcode': 'a', 'name': 'foo'}])
-
-            result = instance._get_categories('a,b')
-            self.assertEqual(result, [{'qcode': 'a', 'name': 'foo'}, {'qcode': 'b', 'name': 'bar'}])
-
-            result = instance._get_categories('a,c')
-            self.assertEqual(result, [{'qcode': 'a', 'name': 'foo'}])
-
-            result = instance._get_categories('c')
-            self.assertEqual(result, [])
-
-            result = instance._get_categories(None)
-            self.assertIsNone(result)
