+++ tests/search/search_test.py
@@ -9,7 +9,7 @@
 # at https://www.sourcefabric.org/superdesk/license
 from eve.utils import ParsedRequest
 
-from . import init_app
+from apps.search import init_app
 from superdesk.metadata.item import ITEM_STATE, CONTENT_STATE
 from superdesk.tests import TestCase
